scala.js -> 1.1.0
Метод js.isUndefined() перестал работать ожидаемым образом.
Прорефакторены всякие is*Available() методы, которые осуществляют
feature-detection.

+++ src1/client/common-sjs/src/main/scala/io/suggest/msg/JsFormatUtil.scala
@@ -59,7 +59,7 @@ object JsFormatUtil {
   }
   implicit object JsDateDayOfWeekHelper extends IDayOfWeekHelper[js.Date] {
     override def apply(d: js.Date): Int = {
-      val d0 = d.getDay()
+      val d0 = d.getDay().toInt
       if (d0 <= 0) 7 else d0
     }
   }
