+++ src1/client/common-sjs/src/main/scala/io/suggest/sjs/dom2/DomQuick.scala
@@ -71,7 +71,7 @@ object DomQuick {
   }
 
   /** Описать текущий сдвиг времени в минутах. */
-  def tzOffsetMinutes: Int = new Date().getTimezoneOffset()
+  def tzOffsetMinutes: Int = new Date().getTimezoneOffset().toInt
 
   /** Текущий год. */
   def currentYear = new Date().getFullYear()
